Move fetch_assembly_metadata to main and refactor accordingly
Move extract_genome_accession to get_assembly_metadata_from_ENA_NCBI,

move species abbreviation generation to fetch_assembly_metadata
