+++ scripts/add_new_species.py
@@ -12,13 +12,10 @@ from pathlib import Path
 from add_new_species.add_content_files import add_assembly_md, add_download_md, add_index_md  # noqa
 from add_new_species.add_stats_file import add_stats_file
 from add_new_species.form_parser import parse_user_form
+from add_new_species.get_assembly_metadata_from_ENA_NCBI import fetch_assembly_metadata
 from add_new_species.image_processer import process_species_image
 from add_new_species.populate_assembly_metadata_fields import populate_assembly_metadata_fields
-from add_new_species.process_data_tracks_Excel import (
-    extract_genome_accession,
-    parse_excel_file,
-    populate_data_tracks_json,
-)
+from add_new_species.process_data_tracks_Excel import parse_excel_file, populate_data_tracks_json
 
 
 def run_argparse() -> argparse.Namespace:
@@ -122,6 +119,11 @@ if __name__ == "__main__":
         sheet_name=args.sheet_name,
     )
 
+    assembly_metadata_dict = fetch_assembly_metadata(
+        data_tracks_list_of_dicts,
+        species_name=user_form_data.species_name,
+    )
+
     add_index_md(
         species_name=user_form_data.species_name,
         species_slug=user_form_data.species_slug,
@@ -158,13 +160,10 @@ if __name__ == "__main__":
     out_img_path = output_dir_paths["image_dir_path"] / f"{user_form_data.species_slug}.webp"
     process_species_image(in_img_path=Path(args.species_image), out_img_path=out_img_path)
 
-    genome_assembly_accession = extract_genome_accession(data_tracks_list_of_dicts)
-
     # WIP note: this function creates and popluates config.yml
     populate_assembly_metadata_fields(
-        accession=genome_assembly_accession,
-        species_name=user_form_data.species_name,
         config_dir_path=output_dir_paths["config_dir_path"],
         content_dir_path=output_dir_paths["content_dir_path"],
         data_tracks_list_of_dicts=data_tracks_list_of_dicts,
+        assembly_metadata_dict=assembly_metadata_dict,
     )
